Conflicts:
	app/View/Layouts/admin_default.ctp

+++ app/Controller/PartnersController.php
@@ -11,12 +11,15 @@ class PartnersController extends AppController {
 			$d = $this->request->data;
 			$d['Partner']['partner_id'] = null;
 			$d['Partner']['partner_date_added'] = date('Ymd');
-			if ($this->Partner->save($d, true, array('partner_id', 'partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page', 'partner_date_added'))) {
-				$this->Session->setFlash("Le partenaire à bien été ajouté !", 'notif');
-				$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
-			} else {
-				$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-				$this->redirect($this->referer());
+			$d['Partner']['partner_img'] = $this->upload_file($d['Partner']['partner_img'], 'partners');
+			if ($d['Partner']['partner_img'] != false) {
+				if ($this->Partner->save($d, true, array('partner_id', 'partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page', 'partner_date_added'))) {
+					$this->Session->setFlash("Le partenaire à bien été ajouté !", 'notif');
+					$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
+				} else {
+					$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
+					$this->redirect($this->referer());
+				}
 			}
 		}
 	}
@@ -27,12 +30,17 @@ class PartnersController extends AppController {
 			$d = $this->request->data;
 			$d['Partner']['partner_id'] = $data['Partner']['partner_id'];
 			$this->Partner->id = $data['Partner']['partner_id'];
-			if ($this->Partner->save($d, true, array('partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page'))) {
-				$this->Session->setFlash("Le partenaire à bien été edité !", 'notif');
-				$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
-			} else {
-				$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-				$this->redirect^($this->referer());
+			$d['Partner']['partner_img'] = $this->upload_file($d['Partner']['partner_img'], 'partners');
+			if ($d['Partner']['partner_img'] != false) {
+				if ($this->delete_file($data['Partner']['partner_img'], 'partners')) {
+					if ($this->Partner->save($d, true, array('partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page'))) {
+						$this->Session->setFlash("Le partenaire à bien été edité !", 'notif');
+						$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
+					} else {
+						$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
+						$this->redirect($this->referer());
+					}
+				}
 			}
 		}
 		$this->request->data = $data;
@@ -40,9 +48,12 @@ class PartnersController extends AppController {
 
 	public function admin_remove($id) {
 		$this->autoRender = false;
-		$this->Partner->delete($id);
-		$this->Session->setFlash("Le partenaire a bien été supprimé !", 'notif');
-		$this->redirect($this->referer());
+		$partner = $this->Partner->find('first', array('conditions' => array('partner_id' => $id)));
+		if ($this->delete_file($partner['Partner']['partner_img'], 'partners')) {
+			$this->Partner->delete($id);
+			$this->Session->setFlash("Le partenaire a bien été supprimé !", 'notif');
+			$this->redirect($this->referer());
+		}
 	}
 
 	public function admin_activated($type, $id) {
